replace SearchModelInfoService with PrepareInfoProtocol
